Factor out MainThreadScheduledExecutorService.
So it can be used by other robolectric tests.
